backend/requirements.txt
aiohttp==3.8.5
python-socketio==5.8.0
numpy==1.24.3
pandas==2.0.2
matplotlib==3.7.1
python-dotenv==1.0.0
pytz==2023.3